+++ CHANGES.md
@@ -1,6 +1,10 @@
-## Version 3.2.0 - February 18, 2020
-  * Code style updates
-  * Fixes to spotbugs findings
+## Version 3.2.0 - TBD
+  * Fix issue with reading WIN file for channels with 0.5 data size
+  * Fix issues adjusting scale using left and right brackets in wave views
+  * Fix helicorder update issue when accessing FDSN WS
+  * Add option to always tile helicorders horizontally (for when there are 4)
+  * Add feature to sort waves by pick times in clipboard
+  * Add feature to align waves by picks in clipboard
 
 ## Version 3.1.0 - February 18, 2020
   * Allow decimals for one bar range and clip value in helicorder settings